+++ pkg/objstore/objtesting/acceptance_e2e_test.go
@@ -49,6 +49,29 @@ func TestObjStore_AcceptanceTest_e2e(t *testing.T) {
 		testutil.Ok(t, err)
 		testutil.Equals(t, "tes", string(content))
 
+		// Unspecified range with offset.
+		rcUnspecifiedLen, err := bkt.GetRange(ctx, "id1/obj_1.some", 1, -1)
+		testutil.Ok(t, err)
+		defer func() { testutil.Ok(t, rcUnspecifiedLen.Close()) }()
+		content, err = ioutil.ReadAll(rcUnspecifiedLen)
+		testutil.Ok(t, err)
+		testutil.Equals(t, "test-data@", string(content))
+
+		// Out of band offset. Do not rely on outcome.
+		// NOTE: For various providers we have different outcome.
+		// * GCS is giving 416 status code
+		// * S3 errors immdiately with invalid range error.
+		// * inmem and filesystem are returning 0 bytes.
+		//rcOffset, err := bkt.GetRange(ctx, "id1/obj_1.some", 124141, 3)
+
+		// Out of band length. We expect to read file fully.
+		rcLength, err := bkt.GetRange(ctx, "id1/obj_1.some", 3, 9999)
+		testutil.Ok(t, err)
+		defer func() { testutil.Ok(t, rcLength.Close()) }()
+		content, err = ioutil.ReadAll(rcLength)
+		testutil.Ok(t, err)
+		testutil.Equals(t, "st-data@", string(content))
+
 		ok, err = bkt.Exists(ctx, "id1/obj_1.some")
 		testutil.Ok(t, err)
 		testutil.Assert(t, ok, "expected exits")
@@ -97,6 +120,7 @@ func TestObjStore_AcceptanceTest_e2e(t *testing.T) {
 		}))
 
 		testutil.Ok(t, bkt.Delete(ctx, "id1/obj_2.some"))
+
 		// Delete is expected to fail on non existing object.
 		// NOTE: Don't rely on this. S3 is not complying with this as GCS is.
 		// testutil.NotOk(t, bkt.Delete(ctx, "id1/obj_2.some"))
@@ -108,5 +132,17 @@ func TestObjStore_AcceptanceTest_e2e(t *testing.T) {
 			return nil
 		}))
 		testutil.Equals(t, []string{"id1/obj_1.some", "id1/obj_3.some"}, seen)
+
+		testutil.Ok(t, bkt.Delete(ctx, "id2/obj_4.some"))
+
+		seen = []string{}
+		testutil.Ok(t, bkt.Iter(ctx, "", func(fn string) error {
+			seen = append(seen, fn)
+			return nil
+		}))
+		expected = []string{"obj_5.some", "id1/"}
+		sort.Strings(expected)
+		sort.Strings(seen)
+		testutil.Equals(t, expected, seen)
 	})
 }