+++ pkg/objstore/objtesting/foreach.go
@@ -1,14 +1,18 @@
 package objtesting
 
 import (
+	"io/ioutil"
 	"os"
 	"testing"
+
+	"github.com/thanos-io/thanos/pkg/objstore/filesystem"
 
 	"github.com/thanos-io/thanos/pkg/objstore"
 	"github.com/thanos-io/thanos/pkg/objstore/azure"
 	"github.com/thanos-io/thanos/pkg/objstore/cos"
 	"github.com/thanos-io/thanos/pkg/objstore/gcs"
 	"github.com/thanos-io/thanos/pkg/objstore/inmem"
+	"github.com/thanos-io/thanos/pkg/objstore/oss"
 	"github.com/thanos-io/thanos/pkg/objstore/s3"
 	"github.com/thanos-io/thanos/pkg/objstore/swift"
 	"github.com/thanos-io/thanos/pkg/testutil"
@@ -25,6 +29,21 @@ func ForeachStore(t *testing.T, testFn func(t testing.TB, bkt objstore.Bucket)) 
 	if ok := t.Run("inmem", func(t *testing.T) {
 		t.Parallel()
 		testFn(t, inmem.NewBucket())
+	}); !ok {
+		return
+	}
+
+	// Mandatory Filesystem.
+	if ok := t.Run("filesystem", func(t *testing.T) {
+		t.Parallel()
+
+		dir, err := ioutil.TempDir("", "filesystem-foreach-store-test")
+		testutil.Ok(t, err)
+		defer testutil.Ok(t, os.RemoveAll(dir))
+
+		b, err := filesystem.NewBucket(dir)
+		testutil.Ok(t, err)
+		testFn(t, b)
 	}); !ok {
 		return
 	}
@@ -113,5 +132,22 @@ func ForeachStore(t *testing.T, testFn func(t testing.TB, bkt objstore.Bucket)) 
 
 	} else {
 		t.Log("THANOS_SKIP_TENCENT_COS_TESTS envvar present. Skipping test against Tencent COS.")
+	}
+
+	// Optional OSS.
+	if _, ok := os.LookupEnv("THANOS_SKIP_ALIYUN_OSS_TESTS"); !ok {
+		bkt, closeFn, err := oss.NewTestBucket(t)
+		testutil.Ok(t, err)
+
+		ok := t.Run("AliYun oss", func(t *testing.T) {
+			testFn(t, bkt)
+		})
+
+		closeFn()
+		if !ok {
+			return
+		}
+	} else {
+		t.Log("THANOS_SKIP_ALIYUN_OSS_TESTS envvar present. Skipping test against AliYun OSS.")
 	}
 }