+++ pkg/server/http/http.go
@@ -1,0 +1,99 @@
+package http
+
+import (
+	"context"
+	"net/http"
+	"net/http/pprof"
+
+	"github.com/go-kit/kit/log"
+	"github.com/go-kit/kit/log/level"
+	"github.com/pkg/errors"
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promhttp"
+	"github.com/thanos-io/thanos/pkg/component"
+	"github.com/thanos-io/thanos/pkg/prober"
+)
+
+// A Server defines parameters for serve HTTP requests, a wrapper around http.Server.
+type Server struct {
+	logger log.Logger
+	comp   component.Component
+	prober *prober.Prober
+
+	mux *http.ServeMux
+	srv *http.Server
+
+	opts options
+}
+
+// New creates a new Server.
+func New(logger log.Logger, reg *prometheus.Registry, comp component.Component, prober *prober.Prober, opts ...Option) *Server {
+	options := options{}
+	for _, o := range opts {
+		o.apply(&options)
+	}
+
+	mux := http.NewServeMux()
+	registerMetrics(mux, reg)
+	registerProfiler(mux)
+	prober.RegisterInMux(mux)
+
+	return &Server{
+		logger: log.With(logger, "service", "http/server", "component", comp.String()),
+		comp:   comp,
+		prober: prober,
+		mux:    mux,
+		srv:    &http.Server{Addr: options.listen, Handler: mux},
+		opts:   options,
+	}
+}
+
+// ListenAndServe listens on the TCP network address and handles requests on incoming connections.
+func (s *Server) ListenAndServe() error {
+	s.prober.SetHealthy()
+	level.Info(s.logger).Log("msg", "listening for requests and metrics", "address", s.opts.listen)
+	return errors.Wrap(s.srv.ListenAndServe(), "serve HTTP and metrics")
+}
+
+// Shutdown gracefully shuts down the server by waiting,
+// for specified amount of time (by gracePeriod) for connections to return to idle and then shut down.
+func (s *Server) Shutdown(err error) {
+	s.prober.SetNotReady(err)
+	defer s.prober.SetNotHealthy(err)
+
+	if err == http.ErrServerClosed {
+		level.Warn(s.logger).Log("msg", "internal server closed unexpectedly")
+		return
+	}
+
+	defer level.Info(s.logger).Log("msg", "internal server shutdown", "err", err)
+
+	if s.opts.gracePeriod == 0 {
+		s.srv.Close()
+		return
+	}
+
+	ctx, cancel := context.WithTimeout(context.Background(), s.opts.gracePeriod)
+	defer cancel()
+
+	if err := s.srv.Shutdown(ctx); err != nil {
+		level.Error(s.logger).Log("msg", "internal server shut down failed", "err", err)
+	}
+}
+
+// Handle registers the handler for the given pattern.
+func (s *Server) Handle(pattern string, handler http.Handler) {
+	s.mux.Handle(pattern, handler)
+}
+
+func registerProfiler(mux *http.ServeMux) {
+	mux.HandleFunc("/debug/pprof/", pprof.Index)
+	mux.HandleFunc("/debug/pprof/cmdline", pprof.Cmdline)
+	mux.HandleFunc("/debug/pprof/profile", pprof.Profile)
+	mux.HandleFunc("/debug/pprof/symbol", pprof.Symbol)
+	mux.HandleFunc("/debug/pprof/trace", pprof.Trace)
+}
+
+func registerMetrics(mux *http.ServeMux, g prometheus.Gatherer) {
+	mux.Handle("/metrics", promhttp.HandlerFor(g, promhttp.HandlerOpts{}))
+}